feat : 모든 템플릿 반응형 뷰 적용
