+++ lib/Screen/ProblemDetail/ProblemDetailScreenWidget.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_tex/flutter_tex.dart';
 import 'package:intl/intl.dart';
+import 'package:ono/Model/TemplateType.dart';
 import 'package:provider/provider.dart';
 
 import '../../GlobalModule/Image/DisplayImage.dart';
@@ -10,6 +11,7 @@ import '../../GlobalModule/Theme/HandWriteText.dart';
 import '../../GlobalModule/Theme/ThemeHandler.dart';
 import '../../GlobalModule/Theme/UnderlinedText.dart';
 import '../../GlobalModule/Util/LatexTextHandler.dart';
+import '../../Model/ProblemModel.dart';
 
 class ProblemDetailScreenWidget{
 
@@ -18,6 +20,101 @@ class ProblemDetailScreenWidget{
     return CustomPaint(
       size: Size.infinite,
       painter: GridPainter(gridColor: themeProvider.primaryColor),
+    );
+  }
+
+  // 공통된 뷰 (풀이 날짜, 문제 출처, 이미지 뷰)
+  Widget buildCommonDetailView(
+      BuildContext context, ProblemModel problemModel, ThemeHandler themeProvider, TemplateType templateType) {
+    final screenWidth = MediaQuery.of(context).size.width;
+    final imageUrl = (templateType == TemplateType.simple) ? problemModel.problemImageUrl : problemModel.processImageUrl;
+
+    if (screenWidth > 600) {
+      // 가로 모드 레이아웃
+      return Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          // 왼쪽 영역 (풀이 날짜와 문제 출처)
+          Flexible(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                const SizedBox(height: 30.0),
+                buildSolvedDate(problemModel.solvedAt, themeProvider),
+                const SizedBox(height: 30.0),
+                buildProblemReference(problemModel.reference, themeProvider),
+              ],
+            ),
+          ),
+          const SizedBox(width: 30.0), // 좌우 간격을 위한 여백
+          // 오른쪽 영역 (이미지 뷰)
+          Flexible(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                const SizedBox(height: 25.0),
+                buildImageSection(context, imageUrl, (templateType == TemplateType.simple) ? '문제 이미지' : '보정 이미지', themeProvider.primaryColor, themeProvider),
+                const SizedBox(height: 30.0),
+              ],
+            ),
+          ),
+        ],
+      );
+    } else {
+      // 세로 모드 레이아웃
+      return Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const SizedBox(height: 16.0),
+          buildSolvedDate(problemModel.solvedAt, themeProvider),
+          const SizedBox(height: 25.0),
+          buildProblemReference(problemModel.reference, themeProvider),
+          const SizedBox(height: 30.0),
+          buildImageSection(context, imageUrl, (templateType == TemplateType.simple) ? '문제 이미지' : '보정 이미지', themeProvider.primaryColor, themeProvider),
+          const SizedBox(height: 30.0),
+        ],
+      );
+    }
+  }
+
+  // 정답 및 풀이 확인 ExpansionTile
+  Widget buildAnalysisExpansionTile(
+      BuildContext context, ProblemModel problemModel, ThemeHandler themeProvider, TemplateType templateType) {
+    final screenWidth = MediaQuery.of(context).size.width;
+    int crossAxisCount = screenWidth > 1100 ? 3 : (screenWidth > 600 ? 2 : 1);
+    double childAspectRatio = screenWidth > 1100 ? 0.8 : 0.9;
+
+    return ExpansionTile(
+      title: Container(
+        padding: const EdgeInsets.all(8.0),
+        child: buildCenteredTitle('해설 및 풀이 확인', themeProvider.primaryColor),
+      ),
+      children: [
+        const SizedBox(height: 10.0),
+        buildSectionWithMemo(problemModel.memo, themeProvider),
+        const SizedBox(height: 20.0),
+        if (templateType == TemplateType.special) buildLatexView(context, problemModel.analysis, themeProvider),
+        const SizedBox(height: 20.0),
+        LayoutBuilder(
+          builder: (BuildContext context, BoxConstraints constraints) {
+            return GridView.count(
+              crossAxisCount: crossAxisCount,
+              shrinkWrap: true,
+              physics: const NeverScrollableScrollPhysics(),
+              mainAxisSpacing: 20.0,
+              crossAxisSpacing: 20.0,
+              childAspectRatio: childAspectRatio,
+              children: (templateType == TemplateType.simple)
+                  ? [_buildImageContainer(context, problemModel.answerImageUrl, '해설 이미지', crossAxisCount, themeProvider)]
+                  : [
+                _buildImageContainer(context, problemModel.problemImageUrl, '원본 이미지', crossAxisCount, themeProvider),
+                _buildImageContainer(context, problemModel.answerImageUrl, '해설 이미지', crossAxisCount, themeProvider),
+                _buildImageContainer(context, problemModel.solveImageUrl, '풀이 이미지', crossAxisCount, themeProvider),
+              ],
+            );
+          },
+        ),
+      ],
     );
   }
 
@@ -119,9 +216,9 @@ class ProblemDetailScreenWidget{
               ],
               child: LatexTextHandler.renderLatex(analysis ?? ""),
               renderingEngine: const TeXViewRenderingEngine.mathjax(),
-              style: TeXViewStyle(
+              style: const TeXViewStyle(
                 elevation: 0,
-                borderRadius: const TeXViewBorderRadius.all(10),
+                borderRadius: TeXViewBorderRadius.all(10),
                 backgroundColor: Colors.white,
               ),
             ),
@@ -147,20 +244,31 @@ class ProblemDetailScreenWidget{
     );
   }
 
-  Widget buildImageSection(
-      BuildContext context,
-      String? imageUrl,
-      String label,
-      Color color,
-      ThemeHandler themeProvider,
-      ) {
-    final mediaQuery = MediaQuery.of(context);
-    double aspectRatio = 0.8; // 기본 비율
+  //
+  static Widget buildCenteredTitle(String text, Color color) {
+    return Container(
+      width: double.infinity,
+      alignment: Alignment.center,
+      padding: const EdgeInsets.symmetric(vertical: 4.0),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          const SizedBox(width: 8.0),
+          UnderlinedText(
+            text: text,
+            fontSize: 30,
+            color: color,
+            fontWeight: FontWeight.bold, // 굵은 텍스트로 설정
+          ),
+        ],
+      ),
+    );
+  }
 
-    // 가로 길이가 1100 이상인 경우 비율을 축소
-    if (mediaQuery.size.width > 1100) {
-      aspectRatio = 1.0; // 원하는 비율로 축소
-    }
+  // 이미지 섹션 빌드 함수
+  Widget buildImageSection(BuildContext context, String? imageUrl, String label, Color color, ThemeHandler themeProvider) {
+    final mediaQuery = MediaQuery.of(context);
+    double aspectRatio = mediaQuery.size.width > 1100 ? 1.0 : 0.8;
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
@@ -182,25 +290,17 @@ class ProblemDetailScreenWidget{
         Center(
           child: GestureDetector(
             onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => FullScreenImage(imagePath: imageUrl),
-                ),
-              );
+              Navigator.push(context, MaterialPageRoute(builder: (context) => FullScreenImage(imagePath: imageUrl)));
             },
             child: Container(
-              width: mediaQuery.size.width * 0.8, // 부모 크기 기준
+              width: mediaQuery.size.width * 0.8,
               decoration: BoxDecoration(
-                color: themeProvider.primaryColor.withOpacity(0.1), // 배경색 추가
-                borderRadius: BorderRadius.circular(10), // 모서리 둥글게 설정
+                color: themeProvider.primaryColor.withOpacity(0.1),
+                borderRadius: BorderRadius.circular(10),
               ),
               child: AspectRatio(
-                aspectRatio: aspectRatio, // 조정된 비율로 이미지의 높이를 조정
-                child: DisplayImage(
-                  imagePath: imageUrl,
-                  fit: BoxFit.contain, // 이미지 전체를 보여주기 위한 설정
-                ),
+                aspectRatio: aspectRatio,
+                child: DisplayImage(imagePath: imageUrl, fit: BoxFit.contain),
               ),
             ),
           ),
@@ -209,24 +309,56 @@ class ProblemDetailScreenWidget{
     );
   }
 
-  //
-  static Widget buildCenteredTitle(String text, Color color) {
-    return Container(
-      width: double.infinity,
-      alignment: Alignment.center,
-      padding: const EdgeInsets.symmetric(vertical: 4.0),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          const SizedBox(width: 8.0),
-          UnderlinedText(
-            text: text,
-            fontSize: 30,
-            color: color,
-            fontWeight: FontWeight.bold, // 굵은 텍스트로 설정
+  // 이미지 컨테이너를 빌드하는 함수
+  Widget _buildImageContainer(BuildContext context, String? imageUrl, String label, int crossAxisCount, ThemeHandler themeProvider) {
+    final mediaQuery = MediaQuery.of(context);
+
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Container(
+          width: double.infinity,
+          alignment: Alignment.centerLeft,
+          padding: const EdgeInsets.symmetric(vertical: 4.0),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.start,
+            children: [
+              Icon(Icons.camera_alt, color: themeProvider.primaryColor),
+              const SizedBox(width: 8.0),
+              HandWriteText(text: label, fontSize: 20, color: themeProvider.primaryColor),
+            ],
           ),
-        ],
-      ),
+        ),
+        const SizedBox(height: 10.0),
+        Expanded(
+          child: Center(
+            child: GestureDetector(
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => FullScreenImage(imagePath: imageUrl),
+                  ),
+                );
+              },
+              child: Container(
+                width: mediaQuery.size.width * 0.8 / crossAxisCount, // 기기 크기 및 그리드 수에 맞게 크기 조절
+                decoration: BoxDecoration(
+                  color: themeProvider.primaryColor.withOpacity(0.1),
+                  borderRadius: BorderRadius.circular(10),
+                ),
+                child: AspectRatio(
+                  aspectRatio: 0.8, // 원하는 비율로 이미지의 높이를 조정
+                  child: DisplayImage(
+                    imagePath: imageUrl,
+                    fit: BoxFit.contain, // 이미지 전체를 보여주기 위한 설정
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ),
+      ],
     );
   }
